Consolidated blogRoute and testimonialsRoute into a single wpRoute
